requirements/common.txt
Django==1.7
django-bootstrap3==4.11.0
psycopg2==2.5.4
Pillow==2.7.0